Reverted to map-based API calls for updates of existing data.
Currently, with Stripe API 9.7.0, there is no way to unset a value by setting null in the builder-pattern API.

Setting to null does not send the value, and the value remains present and unchanged.

Setting to "" results in an error:
You cannot set 'description' to an empty string. We interpret empty strings as null in requests. You may set 'description' to null to delete the property.

We left a boolean toggle.  Perhaps a future version of the Stripe API will resolve this limitation.